Refactored data layer
- Moved networking layer out so that it can be used as a reusable component
- Added mappers from DTO to domain
