+++ CreditScore/CreditScore/App/Data/NetworkDTO/CreditReportDetailsDTO.swift
