+++ CreditScore/CreditScore/App/Data/NetworkEndpoints/BaseURL.swift
@@ -1,0 +1,12 @@
+//
+//  BaseURL.swift
+//  CreditScore
+//
+//  Created by Vivek Jayakumar on 28/1/22.
+//
+
+import Foundation
+
+struct BaseURL {
+    static let clearscoreBaseURL = "5lfoiyb0b3.execute-api.us-west-2.amazonaws.com"
+}
