+++ CreditScore/CreditScore/App/Domain/Entities/CreditScore.swift
@@ -1,0 +1,14 @@
+//
+//  CreditReport.swift
+//  CreditScore
+//
+//  Created by Vivek Jayakumar on 28/1/22.
+//
+
+import Foundation
+
+struct CreditScore {
+    let score: Int
+    let minScore: Int
+    let maxScore: Int
+}
