+++ CreditScore/CreditScore/App/Domain/Use Case/FetchCreditScoreUseCase.swift
@@ -1,0 +1,8 @@
+//
+//  FetchCreditScoreUseCase.swift
+//  CreditScore
+//
+//  Created by Vivek Jayakumar on 28/1/22.
+//
+
+import Foundation
