+++ CreditScore/CreditScore/Networking/Client/EndpointType.swift
@@ -16,25 +16,8 @@ protocol EndpointType {
     var headers: HTTPHeaders? { get set }
     var payload: HTTPParameters? { get set }
     var subPath: String { get }
+    var timeout: Double { get }
     func buildRequest(_ environment: ServerEnvironment) -> URLRequest
 }
 
-extension EndpointType {
-    func pathFor(environment: ServerEnvironment) -> String {
-        switch environment {
-        case .staging:
-            return "/staging/"
-        case .production:
-            return "/prod/"
-        }
-    }
-}
-
-enum HTTPMethod: String {
-    case get = "GET"
-    case post = "POST"
-    case put = "PUT"
-    case patch = "PATCH"
-    case delete = "DELETE"
-}
