+++ CreditScore/CreditScore/Networking/Client/HTTPMethod.swift
@@ -1,0 +1,16 @@
+//
+//  HTTPMethod.swift
+//  CreditScore
+//
+//  Created by Vivek Jayakumar on 28/1/22.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case patch = "PATCH"
+    case delete = "DELETE"
+}
