+++ CreditScore/CreditScore/Networking/Client/ServerEnvironment.swift
@@ -7,10 +7,6 @@
 
 import Foundation
 
-struct BaseURL {
-    static let clearscoreBaseURL = "5lfoiyb0b3.execute-api.us-west-2.amazonaws.com"
-}
-
 enum ServerEnvironment: String {
     case staging = "Staging"
     case production = "Production"
